Updated NTP handling
Realized that using micros() was causing overflow when doing the math
to calculate the offset. Switched to millis() to prevent that (though if
it lasted for 24 *days* then we'd have the same problem).
Also added a 'timing' debug level otherwise too much data is output.
Added States as the NTP radio messages were flooding the network.
Finally, apparently the client_end message value were never set - which
was causing a noteworthy amount of oddness.